refactor: 서비스테스트 추상화
